Add files via upload
This is a Machine learning project. In which I Used data cleaning, EDA techniques, and machine learning models
(Logistic Regression, Linear Discriminant Analysis, Gaussian Naive
Bayes, Decision tree, Random Forest, SVM, and KNN ) to find the
an optimal model that can detect anomalies in the transactions and
detect the cases that might be prone to fraud. 
Using the model selection techniques, finalized Random Forest
the model with 99.96 % accuracy